libtcod.console_new(0, 0) is stupid.
Since the Python wrapper is utterly useless at actually raising errors
(and libtcod itself just returns an "error" value of 0), it was quite
surprising to discover that a certain sub_console wasn't resizable
because it was the root console. Which has id 0. Which is the "error"
value returned from console_new.

Thanks, libtcod. I really needed that.